requirements.txt
django==1.11.*
django-rest-swagger==2.1.2
djangorestframework==3.7.3
django-rest-swagger==2.1.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
Pillow==4.3.0